geodesicDensify is locking up
