+++ src/dymaptic.GeoBlazor.Core/dymaptic.GeoBlazor.Core.csproj
@@ -24,6 +24,7 @@
     </PropertyGroup>
     <PropertyGroup>
         <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+        <DefaultItemExcludes>wwwroot/js/**</DefaultItemExcludes>
     </PropertyGroup>
 
     <ItemGroup>
