+++ tests/dymaptic.GeoBlazor.Core.Test.Blazor/App.razor
