+++ tests/dymaptic.GeoBlazor.Core.Test.Blazor/Pages/Index.razor
@@ -1,0 +1,27 @@
+﻿@page "/"
+
+<h1>Unit Tests</h1>
+<button @onclick="RunTests">Run All Tests</button>
+<div class="result-section">
+	@foreach (var result in _results)
+    {
+        @((MarkupString)result)
+    }
+</div>
+
+<GeometryEngineTests @ref="_geometryEngineTests" OnTestResults="OnTestResults" />
+@code {
+    private async Task RunTests()
+    {
+		_results.Clear();
+        await _geometryEngineTests!.RunTests();
+    }
+
+	private async Task OnTestResults(TestResult result)
+	{
+		_results.Add($"<p><b>{result.ClassName} - <span style=\"color: green;\">Passed: {result.Passed},</span> <span style=\"color: red;\">Failed: {result.Failed}</span></b></p>");
+	}
+
+    private GeometryEngineTests? _geometryEngineTests;
+	private List<string> _results = new List<string>();
+}
