+++ tests/dymaptic.GeoBlazor.Core.Test.Blazor/Pages/_Host.cshtml
@@ -1,6 +1,6 @@
 ﻿@page "/"
 @using Microsoft.AspNetCore.Components.Web
-@namespace dymaptic.GeoBlazor.Tests.Blazor.Pages
+@namespace dymaptic.GeoBlazor.Core.Test.Blazor.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 
 <!DOCTYPE html>
@@ -11,7 +11,7 @@
         <link href="css/site.css" rel="stylesheet" />
          <link href="_content/dymaptic.GeoBlazor.Core"/>
          <link href="_content/dymaptic.GeoBlazor.Core/assets/esri/themes/light/main.css" rel="stylesheet" />
-         <link href="dymaptic.GeoBlazor.Tests.Blazor.styles.css" rel="stylesheet" />
+         <link href="dymaptic.GeoBlazor.Core.Test.Blazor.styles.css" rel="stylesheet" />
         <component type="typeof(HeadOutlet)" render-mode="ServerPrerendered" />
     </head>
     <body>
