+++ tests/dymaptic.GeoBlazor.Core.Test.Blazor/TestResult.cs
@@ -1,0 +1,3 @@
+﻿namespace dymaptic.GeoBlazor.Core.Test.Blazor;
+
+public record TestResult(string ClassName, int Passed, int Failed);
