+++ tests/dymaptic.GeoBlazor.Core.Test.Blazor/_Imports.razor
@@ -11,5 +11,6 @@
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.JSInterop
 @using Microsoft.VisualStudio.TestTools.UnitTesting
-@using dymaptic.GeoBlazor.Tests.Blazor
-@using dymaptic.GeoBlazor.Tests.Blazor.Components
+@using dymaptic.GeoBlazor.Core.Test.Blazor
+@using dymaptic.GeoBlazor.Core.Test.Blazor.Components
+@using TestResult = dymaptic.GeoBlazor.Core.Test.Blazor.TestResult
