+++ tests/dymaptic.GeoBlazor.Core.Test.Blazor/appsettings.Development.json
