--- tests/dymaptic.GeoBlazor.Tests.Blazor/Pages/Index.razor
@@ -1,14 +1,0 @@
-﻿@page "/"
-
-<h1>Unit Tests</h1>
-<button @onclick="RunTests">Run All Tests</button>
-
-<GeometryEngineTests @ref="_geometryEngineTests" />
-@code {
-    private async Task RunTests()
-    {
-        await _geometryEngineTests!.RunTests();
-    }
-
-    private GeometryEngineTests? _geometryEngineTests;
-}
